+++ apps/web/app/contents/random/page.tsx
@@ -1,6 +1,7 @@
 import { BE_URL } from '@/constants';
-import ContentContainer from '@/containers/ContentContainers';
 import { IContent } from '@/types';
+
+import ContentContainer from '@/containers/ContentContainers';
 
 export default async function Page() {
   const contentData = await fetch(`${BE_URL}/api/v1/contents/random?size=1`).then(
