+++ apps/web/app/login/page.tsx
@@ -1,4 +1,4 @@
-import KakaoLogin from "@/containers/Login/KakaoLogin";
+import KakaoLogin from '@/containers/Login/KakaoLogin';
 
 export default function Page() {
   return <KakaoLogin />;
