+++ apps/web/app/page.tsx
@@ -1,6 +1,7 @@
 import { BE_URL } from '@/constants';
-import Home from '@/containers/Home';
 import { IContent } from '@/types';
+
+import Home from '@/containers/Home';
 
 export default async function Page() {
   const contentData = await fetch(`${BE_URL}/api/v1/contents/random?size=5`).then(
