+++ apps/web/containers/Home/index.tsx
@@ -1,24 +1,24 @@
 'use client';
 import { IContents, Sort } from '@/types';
-import styles from './index.module.scss';
 
-import { Swiper, SwiperSlide } from 'swiper/react';
-import { Pagination, Autoplay } from 'swiper/modules';
 import 'swiper/css';
 import 'swiper/css/navigation';
 import 'swiper/css/pagination';
 
+import { useState } from 'react';
+import { Pagination, Autoplay } from 'swiper/modules';
+import { Swiper, SwiperSlide } from 'swiper/react';
+
+import styles from './index.module.scss';
+
+import { FloatBtnBox } from '@/components/ButtonBox/FloatBtnBox';
+import Button from '@/components/Buttons/Button';
+import BannerItem from '@/components/Items/BannerItem';
 import ContentItem from '@/components/Items/ContentItem';
+import { Loading, RoundLoading } from '@/components/Loading';
 import WindowStyle from '@/components/WindowStyles';
 import useInfiniteScroll from '@/hooks/useInfiniteScroll';
 import { getContentsAPI } from '@/services/contents';
-import { useState } from 'react';
-import Loading from '@/components/Loading';
-import RoundLoading from '@/components/Loading/RoundLoading';
-import Button from '@/components/Buttons';
-import { FloatBtnBox } from '@/components/Buttons/FloatBtnBox';
-import BannerItem from '@/components/Items/BannerItem';
-
 interface Props {
   latestContents: IContents[];
 }
@@ -49,7 +49,7 @@ export default function Home({ latestContents }: Props) {
   ) : (
     <div className={styles.wrap}>
       <div className={styles.speechBubbleWrap}>
-        <p>Today's Pick </p>
+        <p>Today&apos;s Pick </p>
         <div className={styles.speechBubblePoint} />
       </div>
       <Swiper
